+++ examples/ce-lb-waap-bot/variables.tf
@@ -1,0 +1,18 @@
+variable "f5xc_api_p12_file" {
+    type = string
+}
+variable "f5xc_api_url" {
+    type = string
+}
+variable "f5xc_tenant"      { 
+    type = string
+}
+variable "f5xc_namespace"   { 
+    type = string
+}
+variable "app_domain"       {
+    type = string
+}
+variable "site_name"        {
+    type = string
+}