+++ examples/re-ce-lb-waap-bot-cdn/1-origin.tf
@@ -1,0 +1,40 @@
+# resource "volterra_healthcheck" "main" {
+#   name      = "http-health-check"
+#   namespace = var.f5xc_namespace
+
+#   http_health_check {
+#     path = "/"
+#   }
+
+#   healthy_threshold   = 1
+#   unhealthy_threshold = 1
+#   interval            = 5
+#   timeout             = 3
+# }
+
+resource "volterra_origin_pool" "ce_juiceshop" {
+  name                   = "ce-juiceshop"
+  namespace = var.f5xc_namespace
+  endpoint_selection     = "LOCAL"
+  loadbalancer_algorithm = "ROUND_ROBIN"
+
+  origin_servers {
+    private_ip {
+      ip        = "10.84.102.251"
+      site_locator {
+        site {
+          namespace = "system"
+          name      = "hv-aws-us-east-1-ce"
+        }
+      }
+    }
+    labels = {}
+  }
+
+  port = 80
+
+  # healthcheck {
+  #   name      = volterra_healthcheck.main.name
+  #   namespace = volterra_healthcheck.main.namespace
+  # }
+}